artisans_Commerce_Cloud/AWS_EC2_Controller/requirements.txt
boto3>=1.26.0
botocore>=1.29.0
dify-plugin>=0.1.0
python-dotenv>=1.0.0
